# alexbrebeanu-IBMpractice-1.0-oas3-resolved.yaml
...
openapi: 3.0.0
info:
  version: 1.0-oas3
servers:
- url: /
paths:
...
  /user/{id}:
    get:
      tags:
      - User
      summary: Find user by ID
      parameters:
      - name: id
        in: path
        description: Id of the user being searched
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: user found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: Invalid userId
    put:
      tags:
      - User
      summary: Update user
      parameters:
      - name: id
        in: path
        description: Id of the user to update
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: confirmed Update
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: ID not found
    delete:
      tags:
      - User
      summary: Delete user
      parameters:
      - name: id
        in: path
        description: Id of user to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: confirmed delete
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: UserId not found
...
components:
  schemas:
    User:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        startYear:
          type: integer
        faculty:
          type: string
        role:
          type: string
          enum:
          - guest
          - student
          - teacher
          - admin
...